# swagger/swaggerDocumentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ium Tweb final assignment api docs",
    "description": "OpenApi swagger docs for final assignment, there are 3 servers and each server has its own routes, here you can find all",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Main Express server containing the frontend and routes that call the other servers, which communicate with the databases."
    },
    {
      "url": "http://localhost:3001/",
      "description": "Express server connected to MongoDB, which stores data that changes frequently."
    },
    {
      "url": "http://localhost:8082/",
      "description": "Spring Boot server communicating with PostgreSQL, containing data that does not change frequently."
    }
  ],
...
  "paths": {
...
    "/importGameLineUp": {
      "post": {
        "tags": ["Main Server"],
        "description": "Route to import game_lineup.csv from the API to the MongoDB database, at the moment path is hard coded with an absolute path",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "File not found error"
          }
        }
      }
    },
    "/importGames": {
      "post": {
        "tags": ["Main Server"],
        "description": "Route to import games.csv from the API to the MongoDB database, at the moment path is hard coded with an absolute path",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "File not found error"
          }
        }
      }
    },
    "/importPlayerValuations": {
      "post": {
        "tags": ["Main Server"],
        "description": "Route to import player_valuations.csv from the API to the MongoDB database, at the moment path is hard coded with an absolute path",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "File not found error"
          }
        }
      }
    },
    "/executeQuery": {
      "post": {
        "tags": ["Main Server"],
        "description": "Route to actually execute a query in the MongoDB database",
        "responses": {
          "200": {
            "description": "OK, query executed successfully"
          },
          "500": {
            "description": "Internal server error, query not executed"
          }
        }
      }
    },
    "/scoreBoard": {
      "get": {
        "tags": ["Main Server"],
        "description": "Route to get the scoreboard from the MongoDB database, it calculate dynamically the points of each team and the position in the league",
        "responses": {
          "200": {
            "description": "OK, scoreboard retrieved"
          },
          "500": {
            "description": "Internal server error, scoreboard not retrieved"
          }
        }
      }
    },
    "/mostValuedPlayers": {
      "get": {
        "tags": ["Main Server"],
        "description": "Route to get the 10 most valued players from the MongoDB database,as in the mongodb collection it call the Spring Boot server to get name",
        "responses": {
          "200": {
            "description": "OK, most valued players retrieved"
          },
          "500": {
            "description": "Internal server error, most valued players not retrieved"
          }
        }
      }
    },
...
  }
}